+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - Update Readme
+- Both flags `--heartbeat` and `--remediation-events` discard any event with status `!= 0`. 
 
 ## [1.0.3] - 2021-04-01
 ### Added